requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
pytest==8.0.0
pytest-django==4.8.0 
setuptools==69.0.2